move StreamType insertion functions to load_metadata
